requirements.txt
feedparser
Pillow
protobuf
python-dotenv
streamlit